# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61ec7076e8f1a8052f4804e2
  title: school
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/exam/{id}":
    delete:
      x-skaffolder-id: 61ec764ee8f1a8052f4805c2
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61ec764ee8f1a8052f4805be
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
    post:
      x-skaffolder-id: 61ec764ee8f1a8052f4805bb
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exam"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 61ec763ee8f1a8052f4805b1
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 61ec7634e8f1a8052f4805a5
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBystudent/{key}":
    get:
      x-skaffolder-id: 61ec7611e8f1a8052f480598
      x-skaffolder-name: findBystudent
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBystudent
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBystudent
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBystudent/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBystudent
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{id}/validate":
    post:
      x-skaffolder-id: 61ec7707e8f1a8052f4805c4
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 61ec7521e8f1a8052f48058b
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: this api used to validate the exam
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{id}/validate
      x-skaffolder-roles: []
      summary: this api used to validate the exam
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    exam:
      x-skaffolder-id: 61ec7521e8f1a8052f48058b
      x-skaffolder-id-db: 61ec7076e8f1a8052f4804e3
      x-skaffolder-id-entity: 61ec7521e8f1a8052f480588
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 61ec7521e8f1a8052f48058a
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 61ec7521e8f1a8052f480589
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 61ec77cde8f1a8052f4805c6
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 61ec7599e8f1a8052f48058e
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61ec7521e8f1a8052f480588
          x-skaffolder-ent2: 61ec74dde8f1a8052f480583
        _teacher:
          x-skaffolder-id: 61ec7599e8f1a8052f48058f
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61ec7521e8f1a8052f480588
          x-skaffolder-ent2: 61ec748fe8f1a8052f48057d
        student:
          x-skaffolder-id: 61ec7599e8f1a8052f48058d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61ec7521e8f1a8052f480588
          x-skaffolder-ent2: 61ec7423e8f1a8052f480576
      required:
        - _id
...